requirements.txt
streamlit
langchain
openai
python-dotenv
langchain_core
pandas
scikit-learn
matplotlib
faiss-cpu
langchain_openai
dotenv
